requirements.txt
pandas
numpy
unidecode
joblib
python-dotenv
python-dateutil
scikit-learn
xgboost
lightgbm
streamlit
fastapi
uvicorn
pydantic
sqlalchemy
pyodbc
beautifulsoup4
selenium
undetected-chromedriver